Prevent filling the jobQueue if jobs are still running
I've seeded 2 million jobs. Everything runs fine running through all the expired jobs. After a period of time the system starts to back up on itself and we hit the recursive locks. There is no reason to keep adding to the queue if there are jobs still being processed. This prevents the system from getting overwhelmed and locked up.